+++ src/App.jsx
@@ -5,9 +5,10 @@ import MainPage from "./pages/MainPage/MainPage"
 import Transaction from "./pages/Transaction/Transaction"
 import BoardPage from "./pages/BoardPage/BoardPage";
 import SignInPage from "./pages/SignInPage/SignInPage";
-import SignUpPage from "./pages/SignUpPage/SignUpPage";
 import PostView from "./components/Board/PostView";
 import PostForm from "./components/Board/PostForm";
+import KakaoCallback from "./components/Auth/SignIn/KakaoCallback";
+import DetailLog from "./components/Auth/SignIn/DetailLog";
 
 function App() {
   return (
@@ -20,7 +21,8 @@ function App() {
           <Route path="/postform" element={<PostForm />} />
           <Route path="/transaction" element={<Transaction />} />
           <Route path="/signin" element={<SignInPage />} />
-          <Route path="/signup" element={<SignUpPage />} />
+          <Route path="/auth" element={<KakaoCallback />} />
+          <Route path="/detaillog" element={<DetailLog />} />
         </Routes>
       </Router>
     </>
