--- src/pages/SignUpPage/SignUpPage.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import SignUp from "../../components/Auth/SignUp/SignUp";
-
-const SignUpPage = () => {
-  return (
-    <>
-      <SignUp />
-    </>
-  );
-};
-
-export default SignUpPage;
